+++ frontend/src/app/main/ui/components/color_input.cljs
@@ -29,6 +29,7 @@
   [props external-ref]
   (let [value     (obj/get props "value")
         on-change (obj/get props "onChange")
+        on-blur   (obj/get props "onBlur")
 
         ;; We need a ref pointing to the input dom element, but the user
         ;; of this component may provide one (that is forwarded here).
@@ -92,6 +93,8 @@
          (mf/deps parse-value apply-value update-input)
          (fn [_]
            (let [new-value (parse-value)]
+             (when on-blur
+               (on-blur))
              (if new-value
                (apply-value new-value)
                (update-input value)))))