+++ frontend/src/app/main/ui/shapes/shape.cljs
@@ -12,7 +12,7 @@
    [app.main.ui.shapes.attrs :as attrs]
    [app.main.ui.shapes.custom-stroke :as cs]
    [app.main.ui.shapes.export :as ed]
-   [app.main.ui.shapes.fill-image :as fim]
+   [app.main.ui.shapes.fills :as fills]
    [app.main.ui.shapes.filters :as filters]
    [app.main.ui.shapes.frame :as frame]
    [app.main.ui.shapes.gradients :as grad]
@@ -63,9 +63,12 @@
       [:defs
        [:& defs/svg-defs          {:shape shape :render-id render-id}]
        [:& filters/filters        {:shape shape :filter-id filter-id}]
-       [:& grad/gradient          {:shape shape :attr :fill-color-gradient}]
        [:& grad/gradient          {:shape shape :attr :stroke-color-gradient}]
-       [:& fim/fill-image-pattern {:shape shape :render-id render-id}]
+       (when (or (some? (:fill-image shape))
+                 (= :image (:type shape))
+                 (> (count (:fills shape)) 1)
+                 (some :fill-color-gradient (:fills shape)))
+         [:& fills/fills            {:shape shape :render-id render-id}])
        [:& cs/stroke-defs         {:shape shape :render-id render-id}]
        [:& frame/frame-clip-def   {:shape shape :render-id render-id}]]
       children]]))