+++ frontend/src/app/main/ui/workspace/sidebar/assets.cljs
@@ -799,7 +799,7 @@
           (let [ids (wsh/lookup-selected @st/state)]
             (if (kbd/alt? event)
               (st/emit! (dc/change-stroke ids color))
-              (st/emit! (dc/change-fill ids color)))))
+              (st/emit! (dc/change-fill ids color 0)))))
 
         rename-color
         (fn [name]